FM: Added the scripts needed for secure boot HookOS (#9)
This creates 2 sets of keys one for UEFI and gpg. These are stored in
setup_scripts as gpg_key and sb_keys.
This creates a signed grub and creates signatures for initramfs and vmlinuz
which will be packaged as a tar file. This tar file will be downloaded as
part of store_alpine tink action.

Signed-off-by: Nagaraj Baglody, Adithya <adithya.nagaraj.baglody@intel.com>