FDE: Enable FDE action.
Pre-requisite before calling FDE, OS needs to be installed and it needs to have a
vfat partition with boot as a label created.

tinker action item to be added in the workflow needs to be as show below.
- name: "setup_fde"
   image: fde:latest
   timeout: 360

Timeout needs to be at least 360. enough time to finish the encryption of the
rootfs and the data that is present in the block device.

Signed-off-by: Nagaraj Baglody, Adithya <adithya.nagaraj.baglody@intel.com>