FM: Client authentication added to hook OS.
This patch creates a new on-boot container which prompts the "SI admin" to
provide a username and password which will be authenticated with the
IDP which is keycloak. The response of the curl https is a JWT which will
be placed in /dev/shm/idp_access_token of the micro OS.
All subsequent containers or actions can access them in their container
from the above location.

To use predefined certificates for client_auth. Place files in client_auth/files
as ca.pem and server_cert.pem.

Signed-off-by: Nagaraj Baglody, Adithya <adithya.nagaraj.baglody@intel.com>