FM: Appending files to initramfs without rebuilding.
Using cpio and gz a final micro OS image is created.
Same folder stucture is needed to complete the build.
run "bash build_image_at_DKAM.sh" for doing a cpio build.

Update all the required files in the etc/idp folder

Note: hook_x86-64.tar.gz needs to be present in the alpine_image folder
Note: grub_source need to at the root of the repo.

These are the prerequisites for running the script.

Signed-off-by: Nagaraj Baglody, Adithya <adithya.nagaraj.baglody@intel.com>