Resize partitions for Virtual Edge Node with DM-Verity support (#109)
Pass ENABLE_DMVERITY_VEN to the fde_dmv action to limit the partition sizes. 
This PR ensures that the Provisioning on Edge Node will not change. 
For VEN we set  rootfs 2048MB,  rootfs hashmap=100MB,
rootfs root hash=50MB, swap=4096MB with reserved=3072MB. 

The remaining size is given to EMT persistent partition or LVMG in the ratio of 40-60.