EMF: DMV: Updated the partitioning needed for EMT. (#157)
EMF: DMV: Updated the partitioning needed for EMT.

This patch includes creation of a reserved partition which is used for
various block disk manipulations needed to expand the filesystem.

EMF: DMV: Fixed a bug in selection of partitioning scheme

This fix ensures that the smaller partitioning will only work on
smaller HDD and not enabled by default.

Signed-off-by: Adithya Baglody <adithya.nagaraj.baglody@intel.com>